+++ BusinessObjects/DTO/BillReqRes/BillResponseDto.cs
@@ -7,13 +7,13 @@ namespace BusinessObjects.DTO.BillReqRes
         public int BillId { get; set; }
         public string? CustomerName { get; set; }
         public string? StaffName { get; set; }
-        public double TotalAmount { get; set; }
-        public double TotalDiscount { get; set; }
+        public decimal TotalAmount { get; set; }
+        public decimal TotalDiscount { get; set; }
         public DateTimeOffset? SaleDate { get; set; }
         public required List<BillItemResponse?> Items { get; set; }
         public required List<BillPromotionResponse?> Promotions { get; set; }
-        public double AdditionalDiscount { get; set; }
+        public decimal AdditionalDiscount { get; set; }
         public int PointsUsed { get; set; }
-        public double FinalAmount { get; set; }
+        public decimal FinalAmount { get; set; }
     }
 }
