+++ BusinessObjects/Models/Promotion.cs
@@ -9,7 +9,7 @@ public partial class Promotion
     public string? Type { get; set; }
     public string? ApproveManager { get; set; }
     public string? Description { get; set; }
-    public double? DiscountRate { get; set; }
+    public decimal? DiscountRate { get; set; }
     public DateTimeOffset? StartDate { get; set; }
     public DateTimeOffset? EndDate { get; set; }
